# swagger_server/models/role.py
# coding: utf-8

from __future__ import absolute_import
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from swagger_server.models.base_model_ import Model
from swagger_server import util


class Role(Model):
    """NOTE: This class is auto generated by the swagger code generator program.

    Do not edit the class manually.
    """
    def __init__(self, id: int=None, name: str=None, description: str=None, active: bool=True):  # noqa: E501
        """Role - a model defined in Swagger

        :param id: The id of this Role.  # noqa: E501
        :type id: int
        :param name: The name of this Role.  # noqa: E501
        :type name: str
        :param description: The description of this Role.  # noqa: E501
        :type description: str
        :param active: The active of this Role.  # noqa: E501
        :type active: bool
        """
        self.swagger_types = {
            'id': int,
            'name': str,
            'description': str,
            'active': bool
        }

        self.attribute_map = {
            'id': 'id',
            'name': 'name',
            'description': 'description',
            'active': 'active'
        }
        self._id = id
        self._name = name
        self._description = description
        self._active = active

    @classmethod
    def from_dict(cls, dikt) -> 'Role':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The Role of this Role.  # noqa: E501
        :rtype: Role
        """
        return util.deserialize_model(dikt, cls)

    @property
    def id(self) -> int:
        """Gets the id of this Role.


        :return: The id of this Role.
        :rtype: int
        """
        return self._id

    @id.setter
    def id(self, id: int):
        """Sets the id of this Role.


        :param id: The id of this Role.
        :type id: int
        """
        if id is None:
            raise ValueError("Invalid value for `id`, must not be `None`")  # noqa: E501

        self._id = id

    @property
    def name(self) -> str:
        """Gets the name of this Role.


        :return: The name of this Role.
        :rtype: str
        """
        return self._name

    @name.setter
    def name(self, name: str):
        """Sets the name of this Role.


        :param name: The name of this Role.
        :type name: str
        """
        if name is None:
            raise ValueError("Invalid value for `name`, must not be `None`")  # noqa: E501

        self._name = name

    @property
    def description(self) -> str:
        """Gets the description of this Role.


        :return: The description of this Role.
        :rtype: str
        """
        return self._description

    @description.setter
    def description(self, description: str):
        """Sets the description of this Role.


        :param description: The description of this Role.
        :type description: str
        """

        self._description = description

    @property
    def active(self) -> bool:
        """Gets the active of this Role.


        :return: The active of this Role.
        :rtype: bool
        """
        return self._active

    @active.setter
    def active(self, active: bool):
        """Sets the active of this Role.


        :param active: The active of this Role.
        :type active: bool
        """

        self._active = active