# swagger_server/models/user_role.py
# coding: utf-8

from __future__ import absolute_import
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from swagger_server.models.base_model_ import Model
from swagger_server import util


class UserRole(Model):
    """NOTE: This class is auto generated by the swagger code generator program.

    Do not edit the class manually.
    """
    def __init__(self, id: int=None, user_id: int=None, role_id: int=None, active: bool=True):  # noqa: E501
        """UserRole - a model defined in Swagger

        :param id: The id of this UserRole.  # noqa: E501
        :type id: int
        :param user_id: The user_id of this UserRole.  # noqa: E501
        :type user_id: int
        :param role_id: The role_id of this UserRole.  # noqa: E501
        :type role_id: int
        :param active: The active of this UserRole.  # noqa: E501
        :type active: bool
        """
        self.swagger_types = {
            'id': int,
            'user_id': int,
            'role_id': int,
            'active': bool
        }

        self.attribute_map = {
            'id': 'id',
            'user_id': 'user_id',
            'role_id': 'role_id',
            'active': 'active'
        }
        self._id = id
        self._user_id = user_id
        self._role_id = role_id
        self._active = active

    @classmethod
    def from_dict(cls, dikt) -> 'UserRole':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The UserRole of this UserRole.  # noqa: E501
        :rtype: UserRole
        """
        return util.deserialize_model(dikt, cls)

    @property
    def id(self) -> int:
        """Gets the id of this UserRole.


        :return: The id of this UserRole.
        :rtype: int
        """
        return self._id

    @id.setter
    def id(self, id: int):
        """Sets the id of this UserRole.


        :param id: The id of this UserRole.
        :type id: int
        """
        if id is None:
            raise ValueError("Invalid value for `id`, must not be `None`")  # noqa: E501

        self._id = id

    @property
    def user_id(self) -> int:
        """Gets the user_id of this UserRole.


        :return: The user_id of this UserRole.
        :rtype: int
        """
        return self._user_id

    @user_id.setter
    def user_id(self, user_id: int):
        """Sets the user_id of this UserRole.


        :param user_id: The user_id of this UserRole.
        :type user_id: int
        """
        if user_id is None:
            raise ValueError("Invalid value for `user_id`, must not be `None`")  # noqa: E501

        self._user_id = user_id

    @property
    def role_id(self) -> int:
        """Gets the role_id of this UserRole.


        :return: The role_id of this UserRole.
        :rtype: int
        """
        return self._role_id

    @role_id.setter
    def role_id(self, role_id: int):
        """Sets the role_id of this UserRole.


        :param role_id: The role_id of this UserRole.
        :type role_id: int
        """
        if role_id is None:
            raise ValueError("Invalid value for `role_id`, must not be `None`")  # noqa: E501

        self._role_id = role_id

    @property
    def active(self) -> bool:
        """Gets the active of this UserRole.


        :return: The active of this UserRole.
        :rtype: bool
        """
        return self._active

    @active.setter
    def active(self, active: bool):
        """Sets the active of this UserRole.


        :param active: The active of this UserRole.
        :type active: bool
        """

        self._active = active