# swagger_server/swagger/swagger.yaml
openapi: 3.0.0
info:
  title: Tantakatu API
  description: This is the first attempt for create API for Tantakatu
  contact:
    name: Gabriel Morales
    email: gabs2323@outlook.com
  license:
    name: Gabriel Morales
  version: 1.0.0
servers:
- url: https://virtserver.swaggerhub.com/Erraticturtle35/Tantakatu/1.0.0
  description: SwaggerHub API Auto Mocking
...
paths:
  /users:
    post:
      tags:
      - users
      summary: Create user
      operationId: create_user
      requestBody:
        description: Created user object
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
        required: true
      responses:
        default:
          description: successful operation
      x-openapi-router-controller: swagger_server.controllers.users_controller
  /users/{username}:
    get:
      tags:
      - users
      summary: Get user by name
      operationId: get_user_by_name
      parameters:
      - name: username
        in: path
        description: The name thats needs to be fetched. Use user1 for testing
        required: true
        style: simple
        explode: false
        schema:
          type: string
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        400:
          description: Invalid username supplied
        404:
          description: User not found
      x-openapi-router-controller: swagger_server.controllers.users_controller
    put:
      tags:
      - users
      summary: Updates users
      description: This is only can be done by the logged in user
      operationId: update_user
      parameters:
      - name: username
        in: path
        description: name that needs to be updated
        required: true
        style: simple
        explode: false
        schema:
          type: string
      requestBody:
        description: Update user object
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
        required: true
      responses:
        400:
          description: Invalid user supplied
        404:
          description: User not found
      x-openapi-router-controller: swagger_server.controllers.users_controller
    delete:
      tags:
      - users
      summary: Delete user
      description: This can only be done by the logged in user.
      operationId: delete_user
      parameters:
      - name: username
        in: path
        description: The name that needs to be deleted
        required: true
        style: simple
        explode: false
        schema:
          type: string
      responses:
        400:
          description: Invalid username supplied
        404:
          description: User not found
      x-openapi-router-controller: swagger_server.controllers.users_controller
  /users/login:
    get:
      tags:
      - users
      summary: Logs user into the system
      operationId: login_user
      parameters:
      - name: username
        in: query
        description: the user name for the login
        required: true
        style: form
        explode: true
        schema:
          type: string
      - name: password
        in: query
        description: The password for the current user
        required: true
        style: form
        explode: true
        schema:
          type: string
      responses:
        200:
          description: Successful operation
          headers:
            X-Rate-Limit:
              description: calls per hour allowed by the user
              style: simple
              explode: false
              schema:
                type: integer
                format: int32
            X-Expires-After:
              description: date in UTC when token expires
              style: simple
              explode: false
              schema:
                type: string
                format: date-time
          content:
            application/json:
              schema:
                type: string
                x-content-type: application/json
        400:
          description: Invalid username/password supplied
      x-openapi-router-controller: swagger_server.controllers.users_controller
...
  /users/{user_id}/roles:
    post:
      tags:
      - users
      summary: Create a role for the user
      operationId: create_user_role
      parameters:
      - name: user_id
        in: path
        description: The user_id thats needs a new role. Use 1 for testing
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          format: int64
      requestBody:
        description: Created user object
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRole'
        required: true
      responses:
        default:
          description: successful operation
      x-openapi-router-controller: swagger_server.controllers.users_controller
  /users/roles/{role_id}:
    get:
      tags:
      - users
      summary: Get role by name
      operationId: get_user_role_by_id
      parameters:
      - name: role_id
        in: path
        description: The id thats needs to be fetched. Use 1 for testing
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          format: int64
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserRole'
        400:
          description: Invalid role id supplied
        404:
          description: role not found
      x-openapi-router-controller: swagger_server.controllers.users_controller
    put:
      tags:
      - users
      summary: Updates role
      description: This is only can be done by the logged in user
      operationId: update_user_role
      parameters:
      - name: role_id
        in: path
        description: id that needs to be updated
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          format: int64
      requestBody:
        description: Update role object
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRole'
        required: true
      responses:
        400:
          description: Invalid role id supplied
        404:
          description: role not found
      x-openapi-router-controller: swagger_server.controllers.users_controller
    delete:
      tags:
      - users
      summary: Delete role
      description: This can only be done by the logged in user.
      operationId: delete_user_role
      parameters:
      - name: role_id
        in: path
        description: The role id that needs to be deleted
        required: true
        style: simple
        explode: false
        schema:
          type: string
      responses:
        400:
          description: Invalid role id supplied
        404:
          description: role not found
      x-openapi-router-controller: swagger_server.controllers.users_controller
  /roles:
    post:
      tags:
      - users
      summary: Create roles
      operationId: create_role
      requestBody:
        description: Created role object
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Role'
        required: true
      responses:
        default:
          description: successful operation
      x-openapi-router-controller: swagger_server.controllers.users_controller
  /roles/{role_id}:
    get:
      tags:
      - users
      summary: Get role by id
      operationId: get_role_by_id
      parameters:
      - name: role_id
        in: path
        description: The id thats needs to be fetched. Use 1 for testing
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          format: int64
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Role'
        400:
          description: Invalid role id supplied
        404:
          description: role not found
      x-openapi-router-controller: swagger_server.controllers.users_controller
    put:
      tags:
      - users
      summary: Updates role
      description: This is only can be done by the logged in user
      operationId: update_role
      parameters:
      - name: role_id
        in: path
        description: id that needs to be updated
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          format: int64
      requestBody:
        description: Update role object
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Role'
        required: true
      responses:
        400:
          description: Invalid role id supplied
        404:
          description: role not found
      x-openapi-router-controller: swagger_server.controllers.users_controller
    delete:
      tags:
      - users
      summary: Delete role
      description: This can only be done by the logged in user.
      operationId: delete_role
      parameters:
      - name: role_id
        in: path
        description: The role id that needs to be deleted
        required: true
        style: simple
        explode: false
        schema:
          type: string
      responses:
        400:
          description: Invalid role id supplied
        404:
          description: role not found
      x-openapi-router-controller: swagger_server.controllers.users_controller
  /roles/permissions/:
    post:
      tags:
      - users
      summary: Create permission
      operationId: create_permission
      requestBody:
        description: Created permission object
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RolePermission'
        required: true
      responses:
        default:
          description: successful operation
      x-openapi-router-controller: swagger_server.controllers.users_controller
  /roles/permissions/{permission_id}:
    get:
      tags:
      - users
      summary: Get permission by id
      operationId: get_permission_by_id
      parameters:
      - name: permission_id
        in: path
        description: The id thats needs to be fetched. Use 1 for testing
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          format: int64
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Permission'
        400:
          description: Invalid permission id supplied
        404:
          description: permission not found
      x-openapi-router-controller: swagger_server.controllers.users_controller
    put:
      tags:
      - users
      summary: Updates permission
      description: This is only can be done by the logged in user
      operationId: update_permission
      parameters:
      - name: permission_id
        in: path
        description: id that needs to be updated
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          format: int64
      requestBody:
        description: Update role object
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Permission'
        required: true
      responses:
        400:
          description: Invalid permission id supplied
        404:
          description: permission not found
      x-openapi-router-controller: swagger_server.controllers.users_controller
    delete:
      tags:
      - users
      summary: Delete permission
      description: This can only be done by the logged in user.
      operationId: delete_permission
      parameters:
      - name: permission_id
        in: path
        description: The permission id that needs to be deleted
        required: true
        style: simple
        explode: false
        schema:
          type: string
      responses:
        400:
          description: Invalid permission id supplied
        404:
          description: permission not found
      x-openapi-router-controller: swagger_server.controllers.users_controller
...
components:
  schemas:
    User:
      required:
      - id
      - login
      - password
      - username
      type: object
      properties:
        id:
          type: integer
          format: int64
        username:
          type: string
        login:
          type: string
        password:
          type: string
        country_id:
          type: string
        active:
          type: boolean
          default: true
      example:
        password: password
        active: true
        id: 0
        login: login
        country_id: country_id
        username: username
...
    UserRole:
      required:
      - id
      - role_id
      - user_id
      type: object
      properties:
        id:
          type: integer
          format: int64
        user_id:
          type: integer
          format: int64
        role_id:
          type: integer
          format: int64
        active:
          type: boolean
          default: true
      example:
        user_id: 6
        role_id: 1
        active: true
        id: 0
    Role:
      required:
      - id
      - name
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        description:
          type: string
        active:
          type: boolean
          default: true
      example:
        name: name
        description: description
        active: true
        id: 0
    RolePermission:
      required:
      - id
      - permission_id
      - role_id
      type: object
      properties:
        id:
          type: integer
          format: int64
        role_id:
          type: integer
          format: int64
        permission_id:
          type: integer
          format: int64
        active:
          type: boolean
          default: true
    Permission:
      required:
      - id
      - name
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        description:
          type: string
        active:
          type: boolean
          default: true
      example:
        name: name
        description: description
        active: true
        id: 0
...